+++ CHANGELOG.md
@@ -29,9 +29,12 @@ Types of changes
 ## [Unreleased] ([git-diff][Unreleased-diff])
 ### Added
  - Dune options file receives `CMAKE_OPTIONS` and `MAKE_OPTIONS` !60
+ - Parameter to use max norm on newton steps !65
+ - Parameter to accept best solution in linear search on newton steps !65
  - Event timestepper !64
 ### Changed
  - TIFF helper is compiled in the dunecopasi library !62
+ - Use `NewtonMethod` instead of `Newton` class from PDELab !65
 ### Fixed
  - Executables can be compiled without the library !60
  - All dune cmake flags are now transitively passed to dune-copasi targets !60